requirements.txt
# Production Dependencies
Flask>=3.0.0
pandas>=2.1.0
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0
Werkzeug>=3.0.0
tensorflow>=2.15.0
scikit-learn>=1.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
